+++ client/app/lib/validator.js
@@ -6,13 +6,15 @@ const { isEmpty } = Ember;
 import presence from 'tahi/lib/validations/presence';
 import number   from 'tahi/lib/validations/number';
 import email    from 'tahi/lib/validations/email';
+import equality from 'tahi/lib/validations/equality';
 
 // 2. Add imported validation to TYPES:
 
 const TYPES = {
   'presence': presence.validation,
   'number': number.validation,
-  'email': email.validation
+  'email': email.validation,
+  'equality': equality.validation
 };
 
 // 3. Add imported validation defaultMessage:
@@ -20,7 +22,8 @@ const TYPES = {
 const DEFAULT_MESSAGES = {
   'presence': presence.defaultMessage,
   'number': number.defaultMessage,
-  'email': email.defaultMessage
+  'email': email.defaultMessage,
+  'equality': equality.defaultMessage
 };
 
 // ---------------------------------