+++ client/tests/addons/add-author-test.js
@@ -1,53 +1,129 @@
-import Ember from "ember";
-import { module, test } from "qunit";
-import startApp from "tahi/tests/helpers/start-app";
-import setupMockServer from '../helpers/mock-server';
-import FactoryGuy from "ember-data-factory-guy";
+import Ember from 'ember';
+import { module, test } from 'qunit';
+import TestHelper from 'ember-data-factory-guy/factory-guy-test-helper';
 import Factory from '../helpers/factory';
-import TestHelper from "ember-data-factory-guy/factory-guy-test-helper";
+import startApp from 'tahi/tests/helpers/start-app';
+import setupMockServer from '../helpers/mock-server';
+import {
+  paperWithTask, addUserAsParticipant, addNestedQuestionToTask
+} from '../helpers/setups';
 
-let App, paper, phase, task, server;
+let App      = null;
+let paper    = null;
+let fakeUser = null;
+let paperId  = null;
+const taskId = 90210;
 
-module("Integration: adding an author", {
+module('Integration: adding an author', {
   afterEach() {
+    server.restore();
     Ember.run(function() { TestHelper.teardown(); });
-    Ember.run(App, "destroy");
+    Ember.run(App, 'destroy');
   },
 
   beforeEach() {
-    App = startApp();
-    server = setupMockServer();
-    TestHelper.setup(App);
+    App      = startApp();
+    server   = setupMockServer();
+    fakeUser = window.currentUserData.user;
 
-    const response = {
+    const records = paperWithTask('AuthorsTask', {
+      id: taskId,
+      oldRole: 'author'
+    });
+
+    Factory.createPermission('AuthorsTask', taskId, ['edit']);
+    TestHelper.handleFindAll('discussion-topic', 1);
+
+    const task = records[1];
+
+    // -- Paper Setup
+
+    const paperPayload = Factory.createPayload('paper');
+    paperId = paperPayload.id;
+    paperPayload.addRecords(records.concat([fakeUser]));
+    const paperResponse = paperPayload.toJSON();
+    paperResponse.participations = [addUserAsParticipant(task, fakeUser)];
+
+    server.respondWith('GET', '/api/papers/' + paperId, [
+      200, {
+        'Content-Type': 'application/json'
+      }, JSON.stringify(paperResponse)
+    ]);
+
+    // -- Task Setup
+
+    const taskPayload = Factory.createPayload('task');
+    taskPayload.addRecords([task, fakeUser]);
+
+    server.respondWith('GET', '/api/tasks/' + taskId, [
+      200, {
+        'Content-Type': 'application/json'
+      }, JSON.stringify(taskPayload.toJSON())
+    ]);
+
+    // -- Nested Question Setup
+
+    const authorResponse = {
       nested_questions: [
-        {id: 120, text: 'A question to be checked', value_type: 'boolean', ident: 'author--published_as_corresponding_author' },
-        {id: 121, text: 'A question to be checked', value_type: 'boolean', ident: 'author--deceased' },
-        {id: 122, text: 'Contributions',            value_type: 'question-set', ident: 'author--contributions' },
-        {id: 123, text: 'A question to be checked', value_type: 'boolean', parent_id: 122, ident: 'author--contributions--conceived_and_designed_experiments' },
-        {id: 124, text: 'A question to be checked', value_type: 'boolean', parent_id: 122, ident: 'author--contributions--performed_the_experiments' },
-        {id: 125, text: 'A question to be checked', value_type: 'boolean', parent_id: 122, ident: 'author--contributions--analyzed_data' },
-        {id: 126, text: 'A question to be checked', value_type: 'boolean', parent_id: 122, ident: 'author--contributions--contributed_tools' },
-        {id: 127, text: 'A question to be checked', value_type: 'boolean', parent_id: 122, ident: 'author--contributions--contributed_writing' },
-        {id: 128, text: 'Other',                    value_type: 'string',  parent_id: 122, ident: 'author--contributions--other' },
+        {id: 1,  text: 'Q', value_type: 'boolean',      ident: 'author--published_as_corresponding_author' },
+        {id: 2,  text: 'Q', value_type: 'boolean',      ident: 'author--deceased' },
+        {id: 3,  text: 'C', value_type: 'question-set', ident: 'author--contributions' },
+        {id: 4,  text: 'Q', value_type: 'boolean',      ident: 'author--contributions--conceptualization' },
+        {id: 5,  text: 'Q', value_type: 'boolean',      ident: 'author--contributions--investigation' },
+        {id: 6,  text: 'Q', value_type: 'boolean',      ident: 'author--contributions--visualization' },
+        {id: 7,  text: 'Q', value_type: 'boolean',      ident: 'author--contributions--methodology' },
+        {id: 8,  text: 'Q', value_type: 'boolean',      ident: 'author--contributions--resources' },
+        {id: 9,  text: 'Q', value_type: 'boolean',      ident: 'author--contributions--supervision' },
+        {id: 10, text: 'Q', value_type: 'boolean',      ident: 'author--contributions--software' },
+        {id: 11, text: 'Q', value_type: 'boolean',      ident: 'author--contributions--data-curation' },
+        {id: 12, text: 'Q', value_type: 'boolean',      ident: 'author--contributions--project-administration' },
+        {id: 13, text: 'Q', value_type: 'boolean',      ident: 'author--contributions--validation' },
+        {id: 14, text: 'Q', value_type: 'boolean',      ident: 'author--contributions--writing-original-draft' },
+        {id: 15, text: 'Q', value_type: 'boolean',      ident: 'author--contributions--writing-review-and-editing' },
+        {id: 16, text: 'Q', value_type: 'boolean',      ident: 'author--contributions--funding-acquisition' },
+        {id: 17, text: 'Q', value_type: 'boolean',      ident: 'author--contributions--formal-analysis' },
+        {id: 18, text: 'Q', value_type: 'boolean',      ident: 'author--government-employee' }
       ]
     };
 
-    server.respondWith('GET', "/api/nested_questions?type=Author", [
-      200, { 'Content-Type': 'application/json' }, JSON.stringify(response)
+    server.respondWith('GET', '/api/nested_questions?type=Author', [
+      200, { 'Content-Type': 'application/json' }, JSON.stringify(authorResponse)
     ]);
 
-    server.respondWith('GET', "/api/admin/journals/authorization", [204, { 'Content-Type': 'application/json' }, "" ]);
-    server.respondWith('GET', "/api/user_flows/authorization", [204, { 'Content-Type': 'application/json' }, "" ]);
-    server.respondWith('GET', "/api/affiliations", [200, { 'Content-Type': 'application/json' }, JSON.stringify([]) ]);
-    server.respondWith('GET', "/api/journals", [200, { 'Content-Type': 'application/json' }, JSON.stringify({journals:[]})]);
+    const taskNestedQuestions = [];
+    taskNestedQuestions.push(Factory.createRecord('NestedQuestion', { id: 19, ident: 'authors--persons_agreed_to_be_named',     text: 'Whatever' }));
+    taskNestedQuestions.push(Factory.createRecord('NestedQuestion', { id: 20, ident: 'authors--authors_confirm_icmje_criteria', text: 'Whatever' }));
+    taskNestedQuestions.push(Factory.createRecord('NestedQuestion', { id: 21, ident: 'authors--authors_agree_to_submission',    text: 'Whatever' }));
+    _.each(taskNestedQuestions, function(q) {
+      addNestedQuestionToTask(q, task);
+    });
 
-    phase = FactoryGuy.make("phase");
-    task = FactoryGuy.make("authors-task", { phase: phase });
-    paper = FactoryGuy.make("paper", { phases: [phase], tasks: [task], editable: true });
-    TestHelper.handleFind(paper);
-    TestHelper.handleFindAll("discussion-topic", 1);
-    Factory.createPermission('AuthorsTask', task.id, ['edit']);
+    server.respondWith('GET', '/api/tasks/' + taskId + '/nested_questions', [
+      200, {
+        'Content-Type': 'application/json'
+      }, JSON.stringify({ nested_questions: taskNestedQuestions })
+    ]);
+
+    server.respondWith('GET', '/api/tasks/' + taskId + '/nested_question_answers', [
+      200, {
+        'Content-Type': 'application/json'
+      }, JSON.stringify({nested_question_answers: []})
+    ]);
+
+    server.respondWith('GET', '/api/admin/journals/authorization', [204, { 'Content-Type': 'application/json' }, '' ]);
+    server.respondWith('GET', '/api/user_flows/authorization', [204, { 'Content-Type': 'application/json' }, '' ]);
+    server.respondWith('GET', '/api/affiliations', [200, { 'Content-Type': 'application/json' }, JSON.stringify([]) ]);
+    server.respondWith('GET', '/api/journals', [
+      200, {
+        'Content-Type': 'application/json'
+      }, JSON.stringify({journals:[]})
+    ]);
+
+    server.respondWith('GET', '/api/countries', [
+      200, {
+        'Content-Type': 'application/json'
+      }, JSON.stringify({countries:['Columbia']})
+    ]);
   }
 });
 
@@ -55,16 +131,31 @@ test('can add a new author', function(assert) {
   Ember.run(function() {
     const firstName = 'James';
 
-    TestHelper.handleFind(task);
     TestHelper.handleCreate('author');
 
-    visit(`/papers/${paper.id}/tasks/${task.id}`);
+    visit(`/papers/${paperId}/tasks/${taskId}`);
     click('.button-primary:contains("Add a New Author")');
     fillIn('.author-first', firstName);
     click('.author-form-buttons .button-secondary:contains("done")');
 
     andThen(function() {
-      assert.ok(find(`.authors-overlay-item .author-name:contains('${firstName}')`).length);
+      assert.ok(find(`.author-task-item .author-name:contains('${firstName}')`).length);
+    });
+  });
+});
+
+test('validation works', function(assert) {
+  Ember.run(function() {
+    TestHelper.handleCreate('author');
+
+    visit(`/papers/${paperId}/tasks/${taskId}`);
+    click('.button-primary:contains("Add a New Author")');
+    click('.author-form-buttons .button-secondary:contains("done")');
+    click('.author-task-item-view-text');
+    click('.author-form-buttons .button-secondary:contains("done")');
+
+    andThen(function() {
+      assert.ok(find('.author-task-item .error').length, 'Errors found');
     });
   });
 });